// src/enums/callbackQueryCommands.enum.ts
enum CallbackQueryCommandsEnum {
    STARSHIP_MOVE_SYSTEM = 'STARSHIP_MOVE_SYSTEM',
    STARSHIP_INVENTORY = 'STARSHIP_INVENTORY',
    PLANET_SCANNER = 'PLANET_SCANNER',
    STARSHIP_MOVE_SYSTEM_CONFIRM = 'MOVE_CONFIRM',
    STARSHIP_INSTALL_TECHNOLOGIES = 'SS_INSTALL_TECH',
    STARSHIP_INSTALL_TECHNOLOGY_CONFIRM = 'SS_INSTALL_CONFIRM',
    CHANGE_COMPANY_NAME = 'CHNG_CMP_NAME',
    FOUND_COLONY = 'FOUND_COLONY',
    FOUND_COLONY_HERE = 'FCH',
    LIST_SELF_COLONIES = 'L_SELF_COL',
    MANAGE_COLONY = 'MANAGE_COL',
    MANAGE_COLONY_MODULES = 'MNG_COL_MOD',
    COLONY_COLLECT_ALL = 'COL_COL_ALL',
    DELIVER_CONSTRUCTION_RESOURCES = 'DEL_CON_RES'
}

export default CallbackQueryCommandsEnum